Refactor Editor help to use BottomSheet.SubSheet component
Avoid duplicative code by leveraging the existing abstraction.
