--- packages/editor/src/components/editor-help/help-detail-navigation-screen.native.js
@@ -1,37 +1,0 @@
-/**
- * External dependencies
- */
-import { View } from 'react-native';
-
-/**
- * WordPress dependencies
- */
-import { BottomSheet } from '@wordpress/components';
-import { useNavigation } from '@react-navigation/native';
-import { __ } from '@wordpress/i18n';
-
-/**
- * Internal dependencies
- */
-import styles from './style.scss';
-
-const HelpDetailNavigationScreen = ( { name, content } ) => {
-	const navigation = useNavigation();
-
-	const goBack = () => {
-		navigation.goBack();
-	};
-
-	return (
-		<BottomSheet.NavigationScreen name={ __( 'What is a block?' ) }>
-			<BottomSheet.NavigationHeader
-				screen={ name }
-				leftButtonOnPress={ goBack }
-			/>
-			<View style={ styles.separator } />
-			{ content }
-		</BottomSheet.NavigationScreen>
-	);
-};
-
-export default HelpDetailNavigationScreen;
